tests/wpt/web-platform-tests/svg/animations/scripted/SVGAnimationElement-exceptions.html

<!DOCTYPE html>
<title>SVGAnimationElement exceptions</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<svg height="0"><animate begin="foo.begin"/></svg>
<script>
setup(function() {
    window.animationElement = document.querySelector('animate');
});

test(function() {
    assert_throws('InvalidStateError', function() { animationElement.getStartTime() });
}, document.title+', getStartTime throws with unresolved interval.');

test(function() {
    assert_throws('NotSupportedError', function() { animationElement.getSimpleDuration() });
}, document.title+', getSimpleDuration throws with undefined simple duration.');
</script>
